refactor: batch 디렉토리 하위 디렉토리 정리
